dance-18 Refactor Dancer class Extract abstract superclass Dancer, add interface with getSteps method and rename.
