+++ src/main/java/com/saponko/model/StepsCountable.java
@@ -1,0 +1,6 @@
+package com.saponko.model;
+
+public interface StepsCountable {
+
+    int getSteps(int[] stepsMap);
+}
